sinov jarayoni tugatildi

+++ WebApi/Service/UpdateUsersBookForAdminService.cs
@@ -1,6 +1,7 @@
 ﻿using Domain.Dto.UserDto;
 using Domain.Entity;
 using Domain.Interface;
+using Microsoft.AspNetCore.Identity;
 using WebApi.Interface;
 
 namespace WebApi.Service;
@@ -8,19 +9,36 @@ namespace WebApi.Service;
 public class UpdateUsersBookForAdminService : IUpdateUsersBookForAdmin<BookDto>
 {
     IRepository<Book> _bookRepository;
+    private UserManager<User> _userManager;
+    private IDetach<Book> _detach;
 
-    public UpdateUsersBookForAdminService(IRepository<Book> bookRepository)
+    public UpdateUsersBookForAdminService(IRepository<Book> bookRepository, UserManager<User> userManager,
+        IDetach<Book> detach)
     {
         _bookRepository = bookRepository;
+        _userManager = userManager;
+        _detach = detach;
     }
 
-    public async Task<BookDto> Create(string userId, BookDto dto)
+    public async Task<ResponceDto> Create(string userId, BookDto dto)
     {
         // ozi bunaqa kitob bomi yomi?
 
-        var book = _bookRepository.GetByIdAsync(dto.BookId);
-        if (book == null)
-            return null;
+        var user = await _userManager.FindByIdAsync(userId);
+        if (user == null)
+            return new ResponceDto()
+            {
+                Message = "Foydalanuvchi topilmadi!",
+                Status = "error 404"
+            };
+
+        if (string.IsNullOrEmpty(dto.Name) || string.IsNullOrEmpty(dto.Author))
+            return new ResponceDto()
+            {
+                Message = "Malumotlar to'ldirilmadi!",
+                Status = "error 400"
+            };
+
         var responce = await _bookRepository.Add(new Book()
         {
             BookId = dto.BookId,
@@ -28,17 +46,19 @@ public class UpdateUsersBookForAdminService : IUpdateUsersBookForAdmin<BookDto>
             Author = dto.Author,
             UserId = userId
         });
-        return new BookDto()
+        return new ResponceDto()
         {
-            BookId = responce.BookId,
-            Name = responce.FullName,
-            Author = responce.Author
+            Message = "Kitob qo'shildi!",
+            Status = "success"
         };
     }
 
 
     public async Task<bool> Delete(int bookId)
     {
+        var book = await _bookRepository.GetByIdAsync(bookId);
+        if (book == null)
+            return false;
         return await _bookRepository.Delete(bookId);
     }
 }
